Unifies point robot
